+++ mininterface/_lib/start.py
@@ -10,7 +10,7 @@ from warnings import warn
 from .._mininterface import Mininterface
 from ..exceptions import DependencyRequired
 from ..interfaces import get_interface
-from ..subcommands import Command, SubcommandPlaceholder
+from ..cli import Command, SubcommandPlaceholder
 from ..tag import Tag
 from .form_dict import DataClass, EnvClass, TagDict, dataclass_to_tagdict