+++ tests/configs.py
@@ -8,7 +8,7 @@ from annotated_types import Gt, Le, Len, Lt
 from tyro.conf import Positional, arg
 
 from mininterface import Tag
-from mininterface.subcommands import Command
+from mininterface.cli import Command
 from mininterface.tag.callback_tag import CallbackTag
 from mininterface.tag.path_tag import PathTag
 from mininterface.tag.alias import Options, Validation